enginel/requirements.txt
django
psycopg[binary]
django-auditlog
django-treebeard
djangorestframework
django-filter
celery[redis]
redis
django-redis
hiredis

# Geometry Processing
cadquery
numpy
scipy

# AWS S3 Storage
django-storages[s3]
boto3

